requirements.txt
rich>=13.7.0
prompt_toolkit>=3.0.43
pandas>=2.0.0
polars>=0.20.0
duckdb>=0.10.0
openpyxl>=3.1.0
xlrd>=2.0.1
chardet>=5.0.0
scipy>=1.13.0
statsmodels>=0.14.0
scikit-learn>=1.4.0
prophet>=1.1.5
lifelines>=0.28.0
mlxtend>=0.23.0
umap-learn>=0.5.0
google-generativeai>=0.7.0
openai>=1.30.0
groq>=0.9.0
httpx>=0.27.0
plotly>=5.20.0
kaleido>=0.2.1
weasyprint>=61.0
jinja2>=3.1.0
keyring>=24.0.0
psutil>=5.9.0
pytest>=8.0.0
pytest-cov>=5.0.0
pyinstaller>=6.0.0
